requirements.txt
absl-py==2.3.1
astunparse==1.6.3
audioread==3.0.1
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
flatbuffers==25.2.10
fonttools==4.60.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.75.0
h5py==3.14.0
idna==3.10
joblib==1.5.2
keras==3.11.3
kiwisolver==1.4.9
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.45.0
Markdown==3.9
markdown-it-py==4.0.0
MarkupSafe==3.0.2
matplotlib==3.10.6
mdurl==0.1.2
ml_dtypes==0.5.3
msgpack==1.1.1
namex==0.1.0
noisereduce==3.0.3
numba==0.62.0
numpy==2.3.3
opt_einsum==3.4.0
optree==0.17.0
packaging==25.0
pillow==11.3.0
platformdirs==4.4.0
pooch==1.8.2
protobuf==6.32.1
pycparser==2.23
Pygments==2.19.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
requests==2.32.5
rich==14.1.0
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
soundfile==0.13.1
soxr==1.0.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.1.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.17.3
